// types.ts
export interface Comment {
  id: string;
  userId: string;
  userName: string;
  text: string;
  createdAt: string; // ISO Date
  likes: number;
  likedBy: string[]; // List of user IDs who liked
  replies?: Comment[]; // Nested replies
}

export interface Article {
  id: string;
  userId?: string; // Author ID (if user generated)
  title: string;
  excerpt: string;
  content: string[]; // Array of paragraphs
  category: 'Filosofia' | 'Benefícios' | 'Inspiração' | 'Anatomia';
  readTime: string;
  imageUrl: string;
  author: string;
  isPremium?: boolean;
  isUserGenerated?: boolean;
  mediaType?: 'image' | 'video' | 'none'; // New field for social posts

  // Social
  likes: number;
  likedBy: string[]; // List of user IDs who liked
  comments: Comment[];
}